Refactor: 컨벤션에 따른 디렉토리 변경 및 코드 수정
